Fix target framework for dependency being ignored.
Fixes issue #42

A NuGet package can specify a dependency on another NuGet package when
a particular target framework is being installed:

    <dependencies>
       <group targetFramework="net40">
          <dependency id="MyDependency" version="1.0"/>
       </group>
    </dependencies>

Installing this NuGet package into a project that targets net20 and
the MyDependency being unavailable on any NuGet feed should not fail.
Only installing the NuGet package into a project that targets net40
will this dependency be resolved.