Fix variables being undefined on next command line in Package Console.
Defining a variable by entering the command line:

    $a = "abc"

Then using the variable again would fail since it would be null.

It seems that the default scoping rules for variables in Pash is
different to Microsoft's PowerShell.

Adding a script via:

    Pipeline.Commands.AddScript(command)

In Pash is equivalent to:

    Pipeline.Commands.AddScript(command, useLocalScope: true);

Whilst in Microsoft's PowerShell this is equivalent to:

    Pipeline.Commands.AddScript(command, useLocalScope: false);